+++ src/crontab_lib/managers/modifiers/__init__.py
@@ -1,1 +1,2 @@
 """Exposed modifiers classes and methods."""
+from crontab_lib.managers.modifiers.crontab_modifier import CrontabModifier