+++ tests/mock_crontab_lib/__init__.py
@@ -1,4 +1,5 @@
 """Exposed mock_crontab_lib classes and methods."""
-from tests.mock_crontab_lib.mock_crontab_manager import mock_crontab_file, mock_entity_cronjob_list, mock_cron_job
+from tests.mock_crontab_lib.mock_crontab_manager import (mock_crontab_file, mock_entity_cron_job_list, mock_cron_job,
+                                                         mock_entity_cron_job)
 
 mock_command_executor_method = "shell_executor_lib.CommandManager.execute_command"